+++ src/app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Category;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -10,18 +11,25 @@ class CategoryController extends Controller
     /**
      * Create a new category.
      *
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @param Request $request
+     * @return RedirectResponse
      */
-    public function create()
+    public function create(Request $request): RedirectResponse
     {
-        request()->validate([
-            'category' => 'required|unique:categories|string',
+        $category = Category::where(['category' => $request->input('category')])
+            ->withTrashed()
+            ->first() ?: new Category();
+
+        $request->validate([
+            'category' => 'required|string|unique:categories,category' . $category->id,
         ]);
 
-        // TODO: If creating a previously deleted category, re-enable it instead of erroring
-        $category = new Category();
-        $category->category = request()->get('category');
-        $category->save();
+        if ($category->trashed()) {
+            $category->restore();
+        } else {
+            $category->category = $request->input('category');
+            $category->save();
+        }
 
         session()->flash('message', 'Category created.');
         return redirect()->back();
@@ -30,11 +38,13 @@ class CategoryController extends Controller
     /**
      * Update a category.
      *
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @param Request $request
+     * @param int $id
+     * @return RedirectResponse
      */
-    public function update(Request $request, int $id)
+    public function update(Request $request, int $id): RedirectResponse
     {
-        request()->validate([
+        $request->validate([
             'category' => 'required|string',
         ]);
 
@@ -49,9 +59,11 @@ class CategoryController extends Controller
     /**
      * Delete a category.
      *
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @param Request $request
+     * @param int $id
+     * @return RedirectResponse
      */
-    public function delete(Request $request, int $id)
+    public function delete(Request $request, int $id): RedirectResponse
     {
         Category::find($id)->delete();